feat(ci/cd): include release toolkit (#213)
This PR introduces the release toolkit for nri-kafka. 
The idea is quite simple, instead of modifying the pipelines we
automatize just the release creation by leveraging the bot. everything
else is kept untouched apart from some renaming.

Notice that we will still be able to create manually a release if
needed, in that case, the changelog should be updated manually